libmpdel: enable use of local UNIX sockets for MPD
By using `make-network-process` we can tune the parameters for either
network connections or local UNIX socket connections.

As an added bonus, connections are very fast.  In my limited testing,
`make-network-process` tends to be dramatically faster than the
`open-network-stream` function.  Of course, without passing `:family
'ipv4` to `make-network-process`, the two functions exhibit similar
performance, so it's not as simple as "one is written in C and the other
is written in ELisp."

Deprecate/obsolete the `libmpdel-port` variable in favor of
`libmpdel-service` which semantically makes more sense now that libmpdel
can connect to either a port or a socket.

`libmpdel-profiles` also work with the new local socket connection
behaviour.

Although I don't use `customize`, the new variable (`libmpdel-service`)
and profiles (`libmpdel-profiles`) should be accessible via `customize`
just the same.

Signed-off-by: Kenny Ballou <kballou@devnulllabs.io>